--- pkg/observe/basic_test.go
@@ -1,12 +1,0 @@
-package observe_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	"iio/pkg/test"
-)
-
-var _ = Describe("basic", func() {
-	test.With("should correctly collect all costs and produce a series", func() {
-
-	})
-})
